refactor(api): Rename package app to api
