+++ api/api.go
@@ -1,4 +1,4 @@
-package app
+package api
 
 import (
 	"database/sql"
